@@ -12,42 +12,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
+    "@angular/animations": "9.1.9",
+    "@angular/common": "9.1.9",
+    "@angular/compiler": "9.1.9",
+    "@angular/core": "9.1.9",
+    "@angular/forms": "9.1.9",
+    "@angular/platform-browser": "9.1.9",
+    "@angular/platform-browser-dynamic": "9.1.9",
+    "@angular/platform-server": "9.1.9",
+    "@angular/router": "9.1.9",
     "@fortawesome/fontawesome-free": "^5.13.0",
     "@nguniversal/module-map-ngfactory-loader": "8.1.1",
     "@types/chart.js": "^2.9.20",
     "angular-bootstrap-md": "^9.2.0",
+    "angularfire2": "^5.4.2",
     "animate.css": "^4.0.0",
     "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.3.1",
+    "bootstrap": "^4.5.0",
     "chart.js": "^2.5.0",
     "core-js": "^3.3.3",
+    "firebase": "^7.14.6",
     "hammerjs": "^2.0.8",
     "jquery": "3.5.0",
     "mdbootstrap": "^4.17.0",
+    "ngx-bootstrap": "^5.6.1",
+    "ngx-spinner": "^9.0.2",
     "oidc-client": "^1.9.1",
     "popper.js": "^1.16.0",
     "rxjs": "^6.5.3",
     "sweetalert2": "^9.10.12",
-    "zone.js": "0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.24",
-    "@angular/cli": "8.3.14",
-    "@angular/compiler-cli": "8.2.12",
-    "@angular/language-service": "8.2.12",
+    "@angular-devkit/build-angular": "~0.901.7",
+    "@angular/cli": "9.1.7",
+    "@angular/compiler-cli": "9.1.9",
+    "@angular/language-service": "9.1.9",
     "@types/jasmine": "~3.4.4",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~12.11.6",
-    "codelyzer": "^5.2.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
@@ -55,7 +60,7 @@
     "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
-    "typescript": "3.5.3"
+    "typescript": "3.8.3"
   },
   "optionalDependencies": {
     "node-sass": "^4.12.0",
