@@ -1,19 +1,34 @@
+// Modulos
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HttpClientModule} from '@angular/common/http';
 import { RouterModule } from '@angular/router';
+import { NgxSpinnerModule } from 'ngx-spinner';
+import { AngularFireModule } from 'angularfire2';
 
+
+// Providers
+import { AngularFireAuth } from 'angularfire2/auth';
+import { AngularFirestore } from 'angularfire2/firestore';
+
+// Pipes
+import { BusquedaPipe } from './pipes/busqueda.pipe';
+
+// Rutas
+import { ROUTES } from './app.routes';
+
+// Componentes
 import { AppComponent } from './app.component';
 import { InicioComponent } from './components/inicio/inicio.component';
 import { MenuComponent } from './components/shared/menu/menu.component';
 import { LoginComponent } from './components/usuario/login/login.component';
 import { SigninComponent } from './components/usuario/signin/signin.component';
-import { ROUTES } from './app.routes';
 import { ProductoComponent } from './components/produtos/producto/producto.component';
 import { RegistrarProductoComponent } from './components/produtos/registrar-producto/registrar-producto.component';
 import { TarjetasComponent } from './components/tarjetas/tarjetas.component';
-import { BusquedaPipe } from './pipes/busqueda.pipe';
+import { environment } from '../environments/environment';
+import { ConfiguracionComponent } from './components/usuario/configuracion/configuracion.component';
 
 
 
@@ -28,16 +43,22 @@ import { BusquedaPipe } from './pipes/busqueda.pipe';
     ProductoComponent,
     RegistrarProductoComponent,
     TarjetasComponent,
-    BusquedaPipe
+    BusquedaPipe,
+    ConfiguracionComponent
   ],
   imports: [
+    AngularFireModule.initializeApp(environment.firebase),
+    BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
+    RouterModule.forRoot(ROUTES, { useHash: false }),
     ReactiveFormsModule,
     HttpClientModule,
     FormsModule,
-    BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
-    RouterModule.forRoot(ROUTES, { useHash: false }) 
+    NgxSpinnerModule,
   ],
-  providers: [],
+  providers: [
+    AngularFireAuth,
+    AngularFirestore
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
