@@ -1,12 +1,14 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace Entity
 {
     public class Producto
     {
+        [Key]
         public string Id { get; set; }
         public string Nombre { get; set; }
-        public string Provedor { get; set; }
+        public string Proveedor { get; set; }
         public int Precio { get; set; }
-        public string Servicio { get; set; }
         public string Presentacion { get; set; }
         public string Descripcion { get; set; }
     }
