@@ -6,13 +6,8 @@ import { Producto } from '../components/models/Producto';
 })
 export class BusquedaPipe implements PipeTransform {
 
-  transform(productos: Producto[], termino: string): any {
-
-      if (termino == null) {
-        return productos;
-      }
-
-      return productos.filter(p => p.nombre.toLowerCase().indexOf(termino.toLowerCase()));
-
+  transform(productos: Producto[], termino: string): Producto[] {
+      if (termino == null) { return productos; }
+      return productos.filter(p => p.nombre.toLowerCase().indexOf(termino.toLowerCase()) !== -1) ;
       }
 }
