@@ -3,7 +3,17 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebase: {
+    apiKey: 'AIzaSyAMZLgGpmSygq2OSAO6xGZUWhhhIcE3F4M',
+    authDomain: 'carniceriafreddy-96400.firebaseapp.com',
+    databaseURL: 'https://carniceriafreddy-96400.firebaseio.com',
+    projectId: 'carniceriafreddy-96400',
+    storageBucket: 'carniceriafreddy-96400.appspot.com',
+    messagingSenderId: '1042519157592',
+    appId: '1:1042519157592:web:396eb6caf99b62b5db932d',
+    measurementId: 'G-6MYQP02VLM'
+  }
 };
 
 /*
